Download invoice and delte product images when update or delete product

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
+const fileHelper  = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   return res.render('admin/edit-product', {
@@ -146,6 +147,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save();
@@ -178,9 +180,15 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(result => {
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('No product found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return  Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
+    .then(() => {
       console.log('Product deleted');
       res.redirect('/admin/products');
     })
